inference/requirements.txt
pyyaml
scikit-learn
xgboost
fastapi
uvicorn[standard]
numpy
pydantic
joblib